pyproject.toml
[project]
name = "tutor"
version = "0.1.0"
description = "tutor using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.14"
dependencies = [
    "crewai[tools]==1.14.1",
    "fpdf2>=2.8.7",
]

[project.scripts]
tutor = "tutor.main:run"
run_crew = "tutor.main:run"
train = "tutor.main:train"
replay = "tutor.main:replay"
test = "tutor.main:test"
run_with_trigger = "tutor.main:run_with_trigger"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
